First split of insights service into dedicated services
