+++ packages/cli/src/modules/insights/insights.collection.service.ts
@@ -1,0 +1,253 @@
+import { Service } from '@n8n/di';
+import { In } from '@n8n/typeorm';
+import { DateTime } from 'luxon';
+import { Logger } from 'n8n-core';
+import type { ExecutionLifecycleHooks } from 'n8n-core';
+import {
+	UnexpectedError,
+	type ExecutionStatus,
+	type IRun,
+	type WorkflowExecuteMode,
+} from 'n8n-workflow';
+
+import { SharedWorkflow } from '@/databases/entities/shared-workflow';
+import { SharedWorkflowRepository } from '@/databases/repositories/shared-workflow.repository';
+import { InsightsMetadata } from '@/modules/insights/database/entities/insights-metadata';
+import { InsightsRaw } from '@/modules/insights/database/entities/insights-raw';
+
+import { InsightsConfig } from './insights.config';
+
+const shouldSkipStatus: Record<ExecutionStatus, boolean> = {
+	success: false,
+	crashed: false,
+	error: false,
+
+	canceled: true,
+	new: true,
+	running: true,
+	unknown: true,
+	waiting: true,
+};
+
+const shouldSkipMode: Record<WorkflowExecuteMode, boolean> = {
+	cli: false,
+	error: false,
+	retry: false,
+	trigger: false,
+	webhook: false,
+	evaluation: false,
+
+	// sub workflows
+	integrated: true,
+
+	// error workflows
+	internal: true,
+
+	manual: true,
+};
+
+type BufferedInsight = Pick<InsightsRaw, 'type' | 'value' | 'timestamp'> & {
+	workflowId: string;
+	workflowName: string;
+};
+
+@Service()
+export class InsightsCollectionService {
+	private readonly cachedMetadata: Map<string, InsightsMetadata> = new Map();
+
+	private bufferedInsights: Set<BufferedInsight> = new Set();
+
+	private flushInsightsRawBufferTimer: NodeJS.Timeout | undefined;
+
+	private isAsynchronouslySavingInsights = true;
+
+	private flushesInProgress: Set<Promise<void>> = new Set();
+
+	constructor(
+		private readonly sharedWorkflowRepository: SharedWorkflowRepository,
+		private readonly insightsConfig: InsightsConfig,
+		private readonly logger: Logger,
+	) {
+		this.logger = this.logger.scoped('insights');
+	}
+
+	startFlushingTimer() {
+		this.isAsynchronouslySavingInsights = true;
+		this.stopFlushingTimer();
+		this.flushInsightsRawBufferTimer = setTimeout(
+			async () => await this.flushEvents(),
+			this.insightsConfig.flushIntervalSeconds * 1000,
+		);
+	}
+
+	stopFlushingTimer() {
+		if (this.flushInsightsRawBufferTimer !== undefined) {
+			clearTimeout(this.flushInsightsRawBufferTimer);
+			this.flushInsightsRawBufferTimer = undefined;
+		}
+	}
+
+	async shutdown() {
+		this.stopFlushingTimer();
+
+		// Prevent new insights from being added to the buffer (and never flushed)
+		// when remaining workflows are handled during shutdown
+		this.isAsynchronouslySavingInsights = false;
+
+		// Wait for all in-progress asynchronous flushes
+		// Flush any remaining events
+		await Promise.all([...this.flushesInProgress, this.flushEvents()]);
+	}
+
+	async workflowExecuteAfterHandler(ctx: ExecutionLifecycleHooks, fullRunData: IRun) {
+		if (shouldSkipStatus[fullRunData.status] || shouldSkipMode[fullRunData.mode]) {
+			return;
+		}
+
+		const status = fullRunData.status === 'success' ? 'success' : 'failure';
+
+		const commonWorkflowData = {
+			workflowId: ctx.workflowData.id,
+			workflowName: ctx.workflowData.name,
+			timestamp: DateTime.utc().toJSDate(),
+		};
+
+		// success or failure event
+		this.bufferedInsights.add({
+			...commonWorkflowData,
+			type: status,
+			value: 1,
+		});
+
+		// run time event
+		if (fullRunData.stoppedAt) {
+			const value = fullRunData.stoppedAt.getTime() - fullRunData.startedAt.getTime();
+			this.bufferedInsights.add({
+				...commonWorkflowData,
+				type: 'runtime_ms',
+				value,
+			});
+		}
+
+		// time saved event
+		if (status === 'success' && ctx.workflowData.settings?.timeSavedPerExecution) {
+			this.bufferedInsights.add({
+				...commonWorkflowData,
+				type: 'time_saved_min',
+				value: ctx.workflowData.settings.timeSavedPerExecution,
+			});
+		}
+
+		if (!this.isAsynchronouslySavingInsights) {
+			// If we are not asynchronously saving insights, we need to flush the events
+			await this.flushEvents();
+		}
+
+		// If the buffer is full, flush the events asynchronously
+		if (this.bufferedInsights.size >= this.insightsConfig.flushBatchSize) {
+			// Fire and forget flush to avoid blocking the workflow execute after handler
+			void this.flushEvents();
+		}
+	}
+
+	private async saveInsightsMetadataAndRaw(insightsRawToInsertBuffer: Set<BufferedInsight>) {
+		const workflowIdNames: Map<string, string> = new Map();
+
+		for (const event of insightsRawToInsertBuffer) {
+			workflowIdNames.set(event.workflowId, event.workflowName);
+		}
+
+		await this.sharedWorkflowRepository.manager.transaction(async (trx) => {
+			const sharedWorkflows = await trx.find(SharedWorkflow, {
+				where: { workflowId: In([...workflowIdNames.keys()]), role: 'workflow:owner' },
+				relations: { project: true },
+			});
+
+			// Upsert metadata for the workflows that are not already in the cache or have
+			// different project or workflow names
+			const metadataToUpsert = sharedWorkflows.reduce((acc, workflow) => {
+				const cachedMetadata = this.cachedMetadata.get(workflow.workflowId);
+				if (
+					!cachedMetadata ||
+					cachedMetadata.projectId !== workflow.projectId ||
+					cachedMetadata.projectName !== workflow.project.name ||
+					cachedMetadata.workflowName !== workflowIdNames.get(workflow.workflowId)
+				) {
+					const metadata = new InsightsMetadata();
+					metadata.projectId = workflow.projectId;
+					metadata.projectName = workflow.project.name;
+					metadata.workflowId = workflow.workflowId;
+					metadata.workflowName = workflowIdNames.get(workflow.workflowId)!;
+
+					acc.push(metadata);
+				}
+				return acc;
+			}, [] as InsightsMetadata[]);
+
+			await trx.upsert(InsightsMetadata, metadataToUpsert, ['workflowId']);
+
+			const upsertMetadata = await trx.findBy(InsightsMetadata, {
+				workflowId: In(metadataToUpsert.map((m) => m.workflowId)),
+			});
+			for (const metadata of upsertMetadata) {
+				this.cachedMetadata.set(metadata.workflowId, metadata);
+			}
+
+			const events: InsightsRaw[] = [];
+			for (const event of insightsRawToInsertBuffer) {
+				const insight = new InsightsRaw();
+				const metadata = this.cachedMetadata.get(event.workflowId);
+				if (!metadata) {
+					// could not find shared workflow for this insight (not supposed to happen)
+					throw new UnexpectedError(
+						`Could not find shared workflow for insight with workflowId ${event.workflowId}`,
+					);
+				}
+				insight.metaId = metadata.metaId;
+				insight.type = event.type;
+				insight.value = event.value;
+				insight.timestamp = event.timestamp;
+
+				events.push(insight);
+			}
+
+			await trx.insert(InsightsRaw, events);
+		});
+	}
+
+	async flushEvents() {
+		// Prevent flushing if there are no events to flush
+		if (this.bufferedInsights.size === 0) {
+			// reschedule the timer to flush again
+			this.startFlushingTimer();
+			return;
+		}
+
+		// Stop timer to prevent concurrent flush from timer
+		this.stopFlushingTimer();
+
+		// Copy the buffer to a new set to avoid concurrent modification
+		// while we are flushing the events
+		const bufferedInsightsToFlush = new Set(this.bufferedInsights);
+		this.bufferedInsights.clear();
+
+		let flushPromise: Promise<void> | undefined = undefined;
+		flushPromise = (async () => {
+			try {
+				await this.saveInsightsMetadataAndRaw(bufferedInsightsToFlush);
+			} catch (e) {
+				this.logger.error('Error while saving insights metadata and raw data', { error: e });
+				for (const event of bufferedInsightsToFlush) {
+					this.bufferedInsights.add(event);
+				}
+			} finally {
+				this.startFlushingTimer();
+				this.flushesInProgress.delete(flushPromise!);
+			}
+		})();
+
+		// Add the flush promise to the set of flushes in progress for shutdown await
+		this.flushesInProgress.add(flushPromise);
+		await flushPromise;
+	}
+}
